Configured docker and ocelot. Added conflicts check
