+++ ChargingStation.Backend/docker-compose.override.yml
@@ -22,10 +22,11 @@ services:
     environment:
       - ASPNETCORE_ENVIRONMENT=Development
       - ASPNETCORE_HTTP_PORTS=8080
-      - "ApiSettings:DepotsServiceAddress=http://depots.api:8080"
+      - "ApiSettings:DepotServiceAddress=http://depots.api:8080"
       - "ApiSettings:ChargePointsServiceAddress=http://chargepoints.api:8080"
       - "ApiSettings:ConnectorsServiceAddress=http://connectors.api:8080"
       - "ApiSettings:WebsocketsServiceAddress=http://chargingstation.websockets:8080"
+      - "ApiSettings:EnergyConsumptionServiceAddress=http://chargingprofiles.api:8080/api/EnergyConsumptionSettings/"
     depends_on:
       - depots.api
       - chargepoints.api
